Adds inheritance to IA object model to reduce duplication
