+++ ia-clientlibrary/src/main/java/org/odpi/egeria/connectors/ibm/ia/clientlibrary/model/NameError.java
@@ -1,0 +1,29 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+/* Copyright Contributors to the ODPi Egeria project. */
+package org.odpi.egeria.connectors.ibm.ia.clientlibrary.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class NameError extends Error {
+
+    @JacksonXmlProperty(isAttribute = true) private String name;
+    @JacksonXmlProperty(isAttribute = true) private ResourceType resourceType;
+
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
+
+    public ResourceType getResourceType() { return resourceType; }
+    public void setResourceType(ResourceType resourceType) { this.resourceType = resourceType; }
+
+    @Override
+    public String toString() {
+        String parent = super.toString();
+        return parent.substring(0, parent.length() - 2)
+                + ", \"name\": \"" + name
+                + "\", \"resourceType\": \"" + resourceType
+                + "\" }";
+    }
+
+}
