+++ ia-clientlibrary/src/main/java/org/odpi/egeria/connectors/ibm/ia/clientlibrary/model/ResourceNameError.java
@@ -1,0 +1,29 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+/* Copyright Contributors to the ODPi Egeria project. */
+package org.odpi.egeria.connectors.ibm.ia.clientlibrary.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class ResourceNameError extends Error {
+
+    @JacksonXmlProperty(isAttribute = true) private String resourceName;
+    @JacksonXmlProperty(isAttribute = true) private ResourceType resourceType;
+
+    public String getResourceName() { return resourceName; }
+    public void setResourceName(String resourceName) { this.resourceName = resourceName; }
+
+    public ResourceType getResourceType() { return resourceType; }
+    public void setResourceType(ResourceType resourceType) { this.resourceType = resourceType; }
+
+    @Override
+    public String toString() {
+        String parent = super.toString();
+        return parent.substring(0, parent.length() - 2)
+                + ", \"resourceName\": \"" + resourceName
+                + "\", \"resourceType\": \"" + resourceType
+                + "\" }";
+    }
+
+}
